testing - single drop
The comparison with EFDC shows things kind of building up in the model domain, which hadn't happened with previous testing of putting a "single drop" of a pollutant (see single-drop.html). I will need to go back into git history to figure out why this stopped happening. This is likely a useful time to return to:

https://github.com/EnvironmentalSystems/ClearWater-riverine/commit/5fe87337536b921b61baf6463beeb78847b79135

https://github.com/EnvironmentalSystems/ClearWater-riverine/commit/37dac52ae0cac548f591f05720d2faeef5bd720b