boundary condition troubleshooting
identified that the issue with weird concentration behavior (high concentration + building up over time) is associated with putting boundary conditions in ghost cells (rather than inside model domain). I think this related to the volumes flowing in (which is used as a proxy for ghost cell volumes) is very high, resulting in very high loads. Need to investigate further.