change order of cell pairs and sign of flow #99
the sign of the face flow was reversed from what's expected physically for the sparse matrix solution, so I changed the sign. The tailwater cell appears to be the cell upstream of the gate and the headwater cell appears to be downstream so i also reversed the order of those in the gate connectivity.